Moved the logic for applying transforms to a converter service in the converters file.
